Merge from CVS: [s48-tuebingen @ 61]
rev 46:  sperber | 2002-03-07 14:03:49 +0100 (Thu, 07 Mar 2002) | 5 lines

Stack continuations have relative offsets different from heap
continuations: the PC is encoded as an offset in stack continuations
and as an address in heap continuations.

rev 52:  sperber | 2002-03-09 18:40:13 +0100 (Sat, 09 Mar 2002) | 4 lines

The OFFSET is *always* just two bytes, as it refers to the CONT-DATA
section in the byte vector.