Subject: Add implementation of FOR-EACH to SRFI-1 [s48-tuebingen @ 508]
This was missing from Olin's reference implementation.

Note that currently, the native versions of MAP and FOR-EACH already
implement the extended semantics specified by SRFI 1.  However, we
might want to enforce the R5RS restrictions to MAP and FOR-EACH at
some time in the future, so it still makes sense to have separate
implementations in SRFI 1.