Subject: Beginnings of Windows port [s48-tuebingen @ 588]
Very rough cut at Windows port.  Basic console operation works, as
well as basic reading and writing to files.  Note done:

- lots of failure cases are ignored
- some known gotchas are marked in the code
- no heavy testing has been done
- I/O for stdin/out/error and files only (no sockets, for example)
- no real build infrastructure
- ...

But it's probably been done in the proper way: It uses overlapped I/O
for files, and threads to get roughly the same effect for the
stdin/out/error.