Subject: New low-level library for dynamically loading external code [s48-tuebingen @ 676]
This is essentially a direct interface to dlopen/dlsym/dlclose.

The difference to the current mechanism in DYNAMIC-EXTERNALS is that
it can map more directly to the layout of current external C code,
doing away with the need for LOOKUP-ALL_EXTERNALS, for instance.
(Albeit, the high-level interface to replace DYXNAMIC-EXTERNALS still
needs to be done.)

Also, the machinery in dynamo.c for keeping track of shared objects
that have already been loaded is much better done in Scheme.

Moreover, I expect to create a Windows version of c/unix/dynlink.c
offering the same interface.