Subject: Correctly identify continuation templates for primitives [s48-tuebingen @ 860]
Before:

> (car 'foo)
1> ,debug
#{Exception-continuation (pc 4) ?}

After:

> (car 'foo)
1> ,debug
#{Exception-continuation (pc 4) car}

We do this by having DEFINE-ALL-OPERATORS also define an ALL-OPERATORS
vector containing all the closed-compiled versions of the primitives,
and have CONTINUATION-TEMPLATE, if all else fails, see if the
continuation it's looking at actually belongs among those.

This avoids the complications of having the closed compilators push
the template, re-arrange the argument stack and so forth.