Subject: Use separate initialization for dynamic externals [s48-tuebingen @ 876]
Before, we would re-load dynamically loaded external libraries on
image resumption using a reinializer.  This is a bad idea because the
resumers might want to call functions in these libraries, and there
are no ordering guarantees among the resumers/reinitializers.
Therefore, introduce a separate initialization phase that runs before
record resumption, and use it for the dynamic externals.