Subject: Revamp I/O error handling [s48-tuebingen @ 882]
Handle errors better in the RTS: Previously, the RTS could get trapped
in continually trying to flush ports on which an error had occurred,
as provisions for error handling were virtually non-existent.  (You
could get 100% CPU usage after an I/O error on a socket, say, because
the buffer flushing would fail, and automatic closing of the port
would also fail because it tried to do buffer flushing first.)

With this patch, the handler for I/O-COMPLETION interrupts also
propagates whether an error occurred along with the message it sends.
We also install an exception handler around the invocations of the
CHANNEL-MAYBE-READ and CHANNEL-MAYBE-WRITE opcodes that causes the
same signalling of an error.  On the other end (in FILL-BUFFER! and
EMPTY-BUFFER!) we perform some treatment preventing further buffer
flushing before raising an exception.

As part of this, partly revert the previous change in the VM, removing
I/O-ERROR interrupts---we still have IO-ERROR-EVENT events, but then
just set a flag in the channel queue to note that the I/O event was
actually an error.

Moreover, actually send out IO_ERROR_EVENT events from the Windows
low-level I/O code.  (These can't occur under Unix, so no work
necessary there.)