Subject: Add VARIABLE-ARGUMENT-LISTS structure [s48-tuebingen @ 969]
I know we dislike optional arguments officially, but occasionally we
need to implement interfaces defined elsewhere, such as those written
by the infamous Dr. S., that have lots of them, and I'm getting tired
of the various kludges floating around.

This structure contains an implementation of PLT's OPT-LAMBDA, which
I've found to be the simplest and most useful piece of infrastructure
for the job.