Subject: Give C code access to enum sets [s48-tuebingen @ 1357]
Previously, the enum-set code would export a bunch of bindings for
callbacks, which was unnecessarily complicated and expensive.  We also
didn't provide C-side functions accessing those bindings.

So---create a new file external-lib.c containing C code that gives
access to Scheme modules, put stuff for the enum sets in there and
splice it all into the build.