Subject: Fix bug in auto-integrator [s48-tuebingen @ 1471]
Report & fix by Taylor Campbell:

The auto-integrator loses with a wrong-type-argument error from the
record accessor for nodes' operator fields if it tries to analyze a
call with the wrong number of arguments to an integrated procedure
when trying to decide whether the caller should be integrated.  [...]

The reason is that the in-line transform returns the node's *form* if
it sees the wrong number of arguments, not the *node* (since it's not
given the node at all), while the analyzer for call nodes compares the
value of the applied in-line transform to the *node*.  The fix is to
make the analyzer for call nodes in scheme/opt/analyze.scm compare the
value to the form instead.