Subject: Change order cont-data pushes [s48-tuebingen @ 1763]
Now push in this order:
template -> environment -> closure
(from environment -> template -> closure)

This is so that environment and closure land in the same slot if only
one of them is included; this makes our job supporting closures in the
byte-code optimizer much simpler.

This also makes the argument order for these three consistent
throughout the compiler and bytecode parser.  (MAKE-FRAME was changed;
everything else remains the same.)