Subject: Don't generate oversize initialization template [s48-tuebingen @ 2148]
The reification code generates a humungous procedure with
straight-line code, that gets compiled into a single large template.
This invisible overflows some assumptions in the segment.  To fix,
split the init code up into procedures of 100 or so, and force the
byte compiler to stick them into a separate template.

Issue #44