Fix deadlock false positive in command-level termination.
Try this:

,open threads locks weak

(let ((lock (make-lock)))
  (spawn
   (lambda ()
     (obtain-lock lock)
     (obtain-lock lock))))

5

,collect
,collect

You get a deadlock.  The reason is that the command-level wait
procedure uses the thread counter to dedcide whether to call the
root-scheduler's wait procedure.  However, that count is inaccurate,
as it includes threads that have already died from GC.  In essence,
the scheduler thinks these threads should still run, but of course
they never do.

To fix, adjust the counter to the actual number of threads terminated
just before scheduling their final deorbit.