Modify build process to generate different files for 32bit and 64bit platforms.
During bootstrapping, autogen.sh generates 32bit and 64bit versions of all
the VM C files (without gc, with twospace, with bibop) and 32bit and 64bit
versions of the initial image.  Configure and make then decide based on
the platform which C file and which initial image to use for making the
final executable and the final image.

TODO: 64bit versions of the various small debugging images.